* Fixed that the middleware for the [Pre/Mid/Post]Dispatch functions was incorrectly sending the subactions to the rest of the middleware chain. (it was sending the raw ActionSet)
Bonus benefit of the above: We don't need the ActionSet unwrapper code in the root reducer.

Another side-effect: Fixes the loading issue in Cypress. (don't know why it happened in Cypress and not the regular online version, but the change fixed it, and was probably due to the react-redux-firebase middleware not having real access to the subactions)

* MS StopBufferingActions() only dispatches an ActionSet action if there are subactions to dispatch. (and if only 1, it just dispatches that subaction itself)